Factor out into debateresponse (individual items inside a debateResponseBlock)
